Make OnRequestLines and OnHeaders belong to WsServer type. refs #8
